+++ container/gxbucketpool/bucketpool.go
@@ -31,7 +31,7 @@ limitations under the License.
  * limitations under the License.
  */
 
-package bucketpool
+package gxbucketpool
 
 import (
 	"math"
