+++ container/gxlru/perf_test.go
@@ -31,5 +31,27 @@ limitations under the License.
  * limitations under the License.
  */
 
-// Package sync2 provides extra functionality along the same lines as sync.
-package sync2
+package gxlru
+
+import (
+	"testing"
+)
+
+type MyValue []byte
+
+func (mv MyValue) Size() int {
+	return cap(mv)
+}
+
+func BenchmarkGet(b *testing.B) {
+	cache := NewLRUCache(64 * 1024 * 1024)
+	value := make(MyValue, 1000)
+	cache.Set("stuff", value)
+	for i := 0; i < b.N; i++ {
+		val, ok := cache.Get("stuff")
+		if !ok {
+			panic("error")
+		}
+		_ = val
+	}
+}
