+++ container/gxsync/consolidator.go
@@ -31,7 +31,7 @@ limitations under the License.
  * limitations under the License.
  */
 
-package sync2
+package gxsync
 
 import (
 	"sync"
@@ -39,7 +39,7 @@ import (
 )
 
 import (
-	"github.com/dubbogo/gost/cache"
+	cache "github.com/dubbogo/gost/container/gxlru"
 )
 
 // Consolidator consolidates duplicate queries from executing simulaneously
@@ -144,13 +144,13 @@ func (cc *ConsolidatorCache) Items() []ConsolidatorCacheItem {
 	return ret
 }
 
-// ccount elements are used with a cache.LRUCache object to track if another
+// ccount elements are used with a gxlru.LRUCache object to track if another
 // request for the same query is already in progress.
 type ccount int64
 
 // Size always returns 1 because we use the cache only to track queries,
 // independent of the number of requests waiting for them.
-// This implements the cache.Value interface.
+// This implements the gxlru.Value interface.
 func (cc *ccount) Size() int {
 	return 1
 }
