+++ container/gxsync/semaphore_flaky_test.go
@@ -31,27 +31,54 @@ limitations under the License.
  * limitations under the License.
  */
 
-package cache
+package gxsync
 
 import (
 	"testing"
+	"time"
 )
 
-type MyValue []byte
-
-func (mv MyValue) Size() int {
-	return cap(mv)
+func TestSemaNoTimeout(t *testing.T) {
+	s := NewSemaphore(1, 0)
+	s.Acquire()
+	released := false
+	go func() {
+		time.Sleep(10 * time.Millisecond)
+		released = true
+		s.Release()
+	}()
+	s.Acquire()
+	if !released {
+		t.Errorf("release: false, want true")
+	}
 }
 
-func BenchmarkGet(b *testing.B) {
-	cache := NewLRUCache(64 * 1024 * 1024)
-	value := make(MyValue, 1000)
-	cache.Set("stuff", value)
-	for i := 0; i < b.N; i++ {
-		val, ok := cache.Get("stuff")
-		if !ok {
-			panic("error")
-		}
-		_ = val
+func TestSemaTimeout(t *testing.T) {
+	s := NewSemaphore(1, 5*time.Millisecond)
+	s.Acquire()
+	go func() {
+		time.Sleep(10 * time.Millisecond)
+		s.Release()
+	}()
+	if s.Acquire() {
+		t.Errorf("Acquire: true, want false")
+	}
+	time.Sleep(10 * time.Millisecond)
+	if !s.Acquire() {
+		t.Errorf("Acquire: false, want true")
+	}
+}
+
+func TestSemaTryAcquire(t *testing.T) {
+	s := NewSemaphore(1, 0)
+	if !s.TryAcquire() {
+		t.Errorf("TryAcquire: false, want true")
+	}
+	if s.TryAcquire() {
+		t.Errorf("TryAcquire: true, want false")
+	}
+	s.Release()
+	if !s.TryAcquire() {
+		t.Errorf("TryAcquire: false, want true")
 	}
 }
